+++ lib/scopo.dart
@@ -3,5 +3,5 @@ export 'src/scope/scope_consumer.dart';
 export 'src/scope/scope_deps.dart';
 export 'src/scope/scope.dart';
 export 'src/progress_iterator/double_progress_iterator.dart';
-export 'src/listenable/listenable.dart';
+export 'src/listenable/listenable_selector.dart';
 export 'src/utils/notifier.dart';