+++ lib/src/navigation_node/navigation_node.dart
@@ -3,6 +3,10 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 /// Navigation node.
+///
+/// A widget that creates a nested `Navigator`. It allows you to include bottom
+/// sheets, dialogs, and other screens in the current scope, ensuring they have
+/// access to all components located above them in the widget tree.
 final class NavigationNode extends StatefulWidget {
   final bool isRoot;
   final Widget child;
@@ -80,6 +84,7 @@ final class _NodeNavigator extends Navigator {
   NavigatorState createState() => NodeNavigatorState();
 }
 
+/// @nodoc
 final class NodeNavigatorState extends NavigatorState {
   Object? _interceptedResult;