+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.2
+
+* breaking changes: rename `ListenableAspectBuilder` to `ListenableSelector`.
+* breaking changes: rename `listenTo` to `select`.
+* update docs.
+
 ## 0.2.0-0.2.1
 
 * breaking changes: remove context from `init`.